Performance fix
This change removes having to create a Dependency object to find out if
another Dependency object matches it. When installing the rails gem,
over 500,000 temporary Dependecy objects (and Requirement and Version)
objects are created to find out if a simple input Depedency object matches.

This is optimized for the simple comparison case of there being no version
requirements on the comparison, but it supports full version comparisons
as well.


git-svn-id: http://rubygems.rubyforge.org/svn/trunk@2474 3d4018f9-ac1a-0410-99e9-8a154d859a19